backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
requests>=2.31.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-text-splitters>=0.3.0
langchain-community>=0.3.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
scikit-learn>=1.5.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.14.0
python-dotenv>=1.0.0
pydantic>=2.10.0
